quiz updated
changed the quiz to only have the 5 questions as decided on stream. I've also put in the basic 'streamer results page' to display the streamers that have been returned from the database. NOTE: the database query is broken, I can't get it to return any values we send it, perhaps the query is very strict and the answers I provide don't satisfy the conditions to retreive any streamers, I don't know. Right now we just retreive the first 5 entries in the DB and send them along as the 'result'